ig2tg_tracker_render_ready/requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.27
python-dotenv>=1.0
httpx>=0.27
psycopg[binary]>=3.1
